+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog
 =========
 
+* Use Brotli quality 4 for streaming responses as well, which is ~100 times faster than the default quality, 11.
+
+  `PR #18 <https://github.com/adamchainz/django-http-compression/pull/18>`__.
+
 1.1.0 (2025-10-15)
 ------------------