+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ Changelog
 
   `PR #18 <https://github.com/adamchainz/django-http-compression/pull/18>`__.
 
+* Decrease minimum length for compression to 50 bytes, copying `CloudFlare’s behaviour <https://developers.cloudflare.com/speed/optimization/content/compression/#:~:text=Minimum%20response%20size%20for%20compression>`__.
+
+  `PR #25 <https://github.com/adamchainz/django-http-compression/pull/25>`__.
+
 * Add a system check warning (``django_http_compression.W001``) when Django’s ``GZipMiddleware`` or `django-compression-middleware <https://pypi.org/project/django-compression-middleware/>`__\’s ``CompressionMiddleware`` are also in use.
 
   `PR #19 <https://github.com/adamchainz/django-http-compression/pull/19>`__.